Swap GetGroupMembers Username and UID assignments
Previously username was assigned to user.Id and UID to user.dn; now it is the inverse.
For GroupManager, the key aspect that matters is that the UID field is matching the PK from the user table (user.id), and for AD while a lot of the times it is the DN it's not always the case.
This was tested via dev build 2.7.0-rc51 (based on commit 10ad0f989a27fad2c7820a68a1b88216af6ec0a6 with a local replace).